Crate modbus

Modbus implementation in pure Rust.

Examples

use modbus::BitValue;
use modbus::tcp::{Ctx, write_single_coil};

// let port = 502;
let mut ctx = Ctx::new_with_port("127.0.0.1", port).unwrap();
assert!(write_single_coil(&mut ctx, 0, BitValue::On).is_ok());

Modules

tcp
The Modbus TCP backend implements a Modbus variant used for communication over TCP/IPv4 networks.

Enums

BitValue
Single bit status values, used in read or write coil functions

ModbusError
Combination of Modbus, IO and data corruption errors

ModbusExceptionCode
Modbus exception codes returned from the server.

Type Definitions

ModbusResult
Result type used to nofify success or failure in communication