Module modbus::tcp

The Modbus TCP backend implements a Modbus variant used for communication over TCP/IPv4 networks.

Structs

Ctx
Context object which holds state for all modbus operations.

Functions

read_coils
Read count bits starting at address addr.

read_discrete_inputs
Read count input bits starting at address addr.

read_holding_registers
Read count 16bit registers starting at address addr.

read_input_registers
Read count 16bit input registers starting at address addr.

write_multiple_coils
Write a multiple coils (bits) starting at address addr.

write_multiple_registers
Write a multiple 16bit registers starting at address addr.

write_single_coil
Write a single coil (bit) to address addr.

write_single_register
Write a single 16bit register to address addr.