Struct modbus::tcp::Transport

pub struct Transport { /* fields omitted */ }

Context object which holds state for all modbus operations.

Methods

impl Transport

pub fn new(addr: &str) -> Result<Transport>

Create a new context context object and connect it to addr on modbus-tcp default port (502)

pub fn new_with_cfg(addr: &str, cfg: Config) -> Result<Transport>

Create a new context object and connect it to addr on port port

pub fn close(&mut self) -> Result<()>