Crate modbus

Modbus implementation in pure Rust.

Examples

use modbus::{Client, Coil};
use modbus::tcp;

let mut cfg = tcp::Config::default();
let mut client = tcp::Transport::new_with_cfg("127.0.0.1", cfg).unwrap();
assert!(client.write_single_coil(0, Coil::On).is_ok());Run

Reexports

pub use tcp::Transport;

pub use tcp::Config;

Modules

scoped
A set of objects which automatically change their register or coil value when they go out of scope

tcp
The Modbus TCP backend implements a Modbus variant used for communication over TCP/IPv4 networks.

Enums

Coil
Single bit status values, used in read or write coil functions

Error
Combination of Modbus, IO and data corruption errors

ExceptionCode
Modbus exception codes returned from the server.

Reason
InvalidData reasons

Traits

Client

Type Definitions

Result
Result type used to nofify success or failure in communication